* 🎨 style(app.vue): add space before self-closing tag This commit adds a space before the self-closing tag of the svgGrid component to improve code readability and consistency.
* ✨ feat(app.vue): add computed property to return all icon objects, refactor search functionality, and add async/await to fetchSvg function
The `allIcons` computed property returns an array of all icon objects in `allData`. The `handleSearch` method now retrieves a lowercase search query string from the target of the event and searches for an icon name in lower case within the `allIcons` array, such that the icon name starts with the search query string. The `submitSearch` method now selects and navigates to the icon found, and scrolls to the icon. The `fetchSvg` function now uses async/await to fetch the SVG and adds the `await` keyword to the `fetch` and `text` methods.

* 🎨 style(tailwind.css): update showcase icon shadow
The showcase icon shadow has been updated to use the `shadow-2xl` class instead of `shadow-md` to improve the visual appearance.

* 🎨 style(svgGrid.vue): reformat code and remove unnecessary whitespace
This commit only reformats the code and removes unnecessary whitespace to improve code readability and consistency.

* 🎨 style(svgGrid.vue): remove unnecessary whitespace and align code
Removed unnecessary whitespace and aligned code to improve readability.

* ♻️ refactor(svgGrid.vue): move mySvgCode to computed property
* 🐛 fix(svgGrid.vue): fix copy function to include svg tags
The mySvgCode function is now moved to a computed property, which is a better practice as it is a function that returns a value. The copy function is fixed to include the svg tags, which was previously missing.