* 🐛 fix(builder.js): change deleteFile function to deleteFiles and delete multiple files * ✨ feat(builder.js): add functionality to output folder structure to a text file * ✨ feat(builder.js): add support for folder-structure.txt file * ✨ feat(builder.js): add support for saving data to a JSON file * ✨ feat(builder.js): add support for deleting multiple files * ✨ feat(builder.js): add support for zipping the svg folder The changes in this commit add several features to the builder.js file. The deleteFile function has been changed to deleteFiles to delete multiple files. The printFolderStructure function has been added to output the folder structure to a text file. The folder-structure.txt file has been added to support this feature. The functionality to save data to a JSON file has been added. The support for deleting multiple files has been added. The support for zipping the svg folder has been added
* 🐛 fix(data.json): fix typo in icon names
The commit fixes a typo in the icon names in the data.json file. The word "strike-through" was changed to "strikethrough" and "subtract" was changed to "substract" to match the correct icon names.

* 🚚 chore(svg): rename strikethrough.svg to strike-through.svg
* 🚚 chore(svg): rename substract.svg to subtract.svg
The commit renames the strikethrough.svg file to strike-through.svg and substract.svg to subtract.svg. This improves the consistency of the naming convention used in the project.

* 🆕 add(folder-structure.txt): add new SVG files to the project
A new folder structure file has been added to the project. It contains a list of SVG files that have been added to the project in the arrows, browser, editor, and finance folders. These SVG files are used to display icons in the application.

* 🔄 chore(refinedicons-all.zip): update icon set
The refinedicons-all.zip file has been updated, likely with new or updated icons.